fix: address audit findings for RCA migration
- C1: Use uint256 for deadline/endsAt in sol struct to match contract's
  EIP-712 typehash (contract struct uses uint64 but typehash string uses
  uint256; ABI encoding is identical)
- C2: Use saturating_add for timestamp arithmetic to prevent overflow
  when duration_seconds defaults to u64::MAX
- W1: Remove redundant field-by-field voucher reconstruction in
  send_indexing_agreement_proposal — use agreement.voucher directly
- W3: Compute `now` timestamp once before the loop instead of per-iteration

+++ dipper-rpc/src/indexer.rs
@@ -50,8 +50,10 @@ pub mod indexer_client {
             /// Matches `IRecurringCollector.RecurringCollectionAgreement` exactly.
             struct RecurringCollectionAgreement {
                 bytes16 agreementId;
-                uint64 deadline;
-                uint64 endsAt;
+                // NB: The on-chain struct declares these as uint64 for storage efficiency,
+                // but the EIP-712 typehash uses uint256. We must match the typehash.
+                uint256 deadline;
+                uint256 endsAt;
                 address payer;
                 address dataService;
                 address serviceProvider;
